requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
flask==3.1.1
greenlet==3.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
packaging==25.0
python-dotenv==1.1.0
requests==2.32.3
sqlalchemy==2.0.41
typing-extensions==4.13.2
urllib3==2.4.0
werkzeug==3.1.3
